+++ DelvUI/Interface/Party/PartyFramesBar.cs
@@ -42,6 +42,8 @@ namespace DelvUI.Interface.Party
 
         private SmoothHPHelper _smoothHPHelper = new SmoothHPHelper();
 
+        private bool _wasHovering = false;
+
         public IPartyFramesMember? Member;
 
         public PartyFramesBar(
@@ -135,21 +137,36 @@ namespace DelvUI.Interface.Party
             {
                 MouseOverHelper.Instance.Target = character;
 
-                // move player bar to this spot on ctrl+alt+shift click
-                if (ImGui.GetIO().KeyCtrl && ImGui.GetIO().KeyAlt && ImGui.GetIO().KeyShift && ImGui.GetIO().MouseClicked[0])
+                if (!_wasHovering)
                 {
-                    MovePlayerEvent?.Invoke(this);
+                    MouseOverHelper.Instance.StartHandlingMouseInputs();
+                    _wasHovering = true;
                 }
-                // target
-                else if (ImGui.IsMouseClicked(ImGuiMouseButton.Left) && character != null)
+
+                // left click
+                if (MouseOverHelper.Instance.LeftButtonClicked)
                 {
-                    Plugin.TargetManager.SetTarget(character);
+                    // move player bar to this spot on ctrl+alt+shift click
+                    if (ImGui.GetIO().KeyCtrl && ImGui.GetIO().KeyAlt && ImGui.GetIO().KeyShift)
+                    {
+                        MovePlayerEvent?.Invoke(this);
+                    }
+                    // target
+                    else if (character != null)
+                    {
+                        Plugin.TargetManager.SetTarget(character);
+                    }
                 }
-                // context menu
-                else if (ImGui.IsMouseClicked(ImGuiMouseButton.Right))
+                // right click (context menu)
+                else if (MouseOverHelper.Instance.RightButtonClicked)
                 {
                     OpenContextMenuEvent?.Invoke(this);
                 }
+            }
+            else if (_wasHovering)
+            {
+                MouseOverHelper.Instance.StopHandlingMouseInputs();
+                _wasHovering = false;
             }
 
             // bg