+++ DelvUI/changelog.md
@@ -1,5 +1,7 @@
 # 0.3.0.2
 - Fixed smooth HP in party frames not being smooth enough.
+- Fixed mouseover not working for actions bound to mouse buttons (ie M3, M4, etc).
+- Added a mouseover setting under Misc > HUD Options to enable and disable the feature completely.
 
 # 0.3.0.1
 - Fixed smooth HP in party frames.